fix: change age column type to float in convert_cols function and update corresponding test

+++ src/data/cleaning.py
@@ -72,7 +72,7 @@ def convert_cols(df):
     df['Age recode with single ages and 90+'] = (
     df['Age recode with single ages and 90+']
     .str.extract(r'(\d+)', expand=False)
-    .astype(int))
+    .astype(float))
     return df
 
 def event_flag(df):
